+++ elasticdl/pkg/ps/parameter.go
@@ -1,6 +1,6 @@
 package ps
 
-import "elasticdl.org/elasticdl/common/tensor"
+import "elasticdl.org/elasticdl/pkg/common"
 
 // Parameter contains non-embedding param and embedding param
 type Parameter struct {
