+++ elasticdl/pkg/ps/server_test.go
@@ -1,4 +1,4 @@
-package main
+package ps
 
 import "testing"
 import "os"
